refactor(router): extract auth-guard + lazy helpers to guards.ts (#89)
Adding the matrix route names pushed routes.ts to 620 lines, over its 618
baseline and the 600 ceiling (code-quality-audit failed in CI). Extract the
cross-cutting createAuthGuard factory, lazyRouteComponent, and the
import.meta.glob maps into a cohesive router/guards.ts. routes.ts is now 581
lines (under the ceiling) and its baseline entry is removed.

No behavior change: helpers re-imported from './guards'. type-check, eslint,
and the router specs (matrix-nav + nddscore) pass.
